pyproject.toml
[project]
name = "job-search-agent"
version = "0.1.0"
description = "Periodically scans job boards for matching roles and sends email digests"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
